Parallel runtime thread pinning.
This change introduces two new features in the mercury runtime;
pinning of threads to CPU cores/threads and runtime detection of the number of
CPU cores/threads available.

If MR_num_threads has not been specified in the runtime options with the -P
flag we use the sysconf(_SC_NPROCESSORS_ONLN) call if available to detect the
number of CPUs online and set MR_num_threads available.  As before this
defaults to 1.

Thread pinning is enabled if the runtime was able to detect the number of CPUs
on the machine or the user specifically requests thread pinning with the
--thread-pinning runtime option.  The sched_setaffinity() call is used to pin
each thread to a specific CPU.

I believe that in some cases thread pinning can achieve better performance,
this is yet to be determined and it may depend on the machine's architecture.
It does make profiling of the runtime system more reliable where the RDTSCP
instruction is not available.  It ensuring that a thread is not migrated to a
different CPU between sampling of the CPU's TSC.

configure.in:
runtime/mercury_conf.h.in:
	Detect the presence of sched.h sysconf() sched_setaffinity() and
	_SC_NPROCESSORS_ONLN.

doc/user_guide.texi:
	Document the new --thread-pinning runtime option.
	Adjust the documentation of -P to reflect the new behaviour.

runtime/mercury_context.c:
	Add the MR_pin_thread() function.
	Create a new global MR_bool MR_pin_threads;
	Add the calculation of the number of threads to use to
	MR_init_thread_stuff()
	Correct a bug in a format string in my previous patch.

runtime/mercury_context.h:
	Export the new MR_pin_thread() function.
	Export the new MR_pin_threads global.
	Correct a previous spelling mistake.
	Adjust the documentation of MR_init_thread_stuff to reflect the new
	behaviour.

runtime/mercury_wrapper.c:
	Pin the primordial thread to a CPU after it spawns the other threads.
	Add the --thread-pinning runtime configuration option.
	Move the calculation of MR_max_outstanding_contexts until after
	MR_init_thread_stuff() so that it is calculated after the number of CPUs
	available has been determined.
	Add a pause instruction to a spinloop for better behaviour on later
	i386/x86_64 processors.  See the documentation for MR_ATOMIC_PAUSE.

runtime/mercury_thread.c:
	After a thread is spawned call MR_pin_thread() to pin a thread to a CPU if
	the thread has been created to pickup work from the global work queue.