Implement a command within the term browser that prints the representation
Estimated hours taken: 3
Branches: main

Implement a command within the term browser that prints the representation
of the selected term. The command is named "addr" or "memory_addr", since
it adds new functionality only if the term is a possibly tagged pointer.
(If it is an integer or character, a plain "print" command would already
do the job.) This is intended mainly for Mercury system developers.

Note that this diff does not add documentation of the new browser command,
because I cannot find anyplace the existing browser commands are documented,
so I do not know where to add the documentation to.

browser/parse.m:
	Put the browser command types in a logical order, with related commands
	being together.

	Make the code that recognizes browser command types have the same order
	as the definition of the browser command type.

	Add code to recognize the new command for the new functionality.

	Add prefixes to the function symbols of the command and token types
	to avoid ambiguities, and avoid using graphic characters that need to
	be quoted.

browser/browser_info.m:
	Add prefixes to the function symbols of the debugger type
	to avoid using the keyword "external" as a function symbol.

browser/browse.m:
	Make the switch on the browser command type have the same order as the
	definition of the browser command type.

	Add code to implement the new command.

browser/declarative_user.m:
	Conform to the changes above.