Remove broken parallel execution profiling code.
Parallel execution profiling can be enabled by defining
MR_PROFILE_PARALLEL_EXECUTION_SUPPORT when building the runtime, libraries and
applications.  This collects information about the performance of various
parallel execution operations such as the management of contexts and sparks.

in the case of sparks this was broken.  It had gone unnoticed since it was only
compiled in if the MR_join_and_continue macro was evaluated while
MR_PROFILE_PARALLEL_EXECUTION_SUPPORT was defined.  The C compiler would be
unable to compile any program making use of parallel conjunctions.

This commit removes the broken code, I plan to fix it and re-add it later.

runtime/mercury_context.h:
    As above.