Address bug #15.  In the current implementation if you have
   ( if some [V] C then T else E )

then V will be quantified over both C and T.  This was not
documented anywhere.  The fix is to document it.

(There has been some discussion in the notes for bug #15 about
adding new syntax to explicitly support the situation where you
want to quantify over both the condition and then branch of
and if-then-else.  The situation occurs so rarely that it's
probably not worth it - also, the current proposals don't work
with the ( C -> T ; E ) syntax for if-then-elses.)

doc/reference_manual.texi:
	Document that the scope of an existential quantification that
	occurs as the top-level goal of the condition of an if-then-else
	extends over the then branch.

	Add a couple of illustrative examples involving existential
	quantifications in if-then-else conditions.

doc/faq.texi:
	Delete part of an answer that is causing confusion.
	This doesn't have any impact on the rest of the answer and avoids
	the potential confusion.