Re-order independent parallel conjunctions to prevent a pathological case of memory usage.
The right hand side of the parallel conjunction operator is potentially forked
off for parallel execution while the left hand side is executed immediately by
the current context.  If the left hand side finishes first the context must be
suspended, with it's stack saved since this stack contains callers' stack
frames that will be needed on the completion of the current procedure.  When
the left operand is simple and the right operand is recursive this creates a
linear number of stacks that must be saved and the OS quickly runs out of
memory.  This patch re-orders independent conjunctions so that the recursive
case is executed as the left operand and the simple case is the right operand.
This also has the benefit of creating all the sparks involved with the
computation very quickly.

This transformation has been added to the dependant parallel conjunction
transformation module since it fits nicely when the dependant parallel
conjunction transformation encounters an independent parallel conjunction.

compiler/dep_par_conj.m:
    As above.

    Modified the sync_info structure used in the dependant parallel conjunction
    transformation to track the pred_proc_id of the current procedure.

    Call the transformation for re-ordering independent parallel conjuncts
    during the dependant parallel conjunction transformation for independent
    conjunctions.