There is a race condition between the initialisation of modules
Estimated hours taken: 4
Branches: main

There is a race condition between the initialisation of modules
and mutables, mainly because there is no guaranteed order of
initialisation.  This is a problem for pretty_printer.m since
we want other modules to be able to register formatters
for types they define via an ':- initialise' directive.

This patch fixes the problem by using foreign code instead of a
mutable for the default formatter map variable.