Re-throw unhandled signals.
The Mercury runtime uses signal handlers to implement redzones and to provide
extra information in the case of a segmentation violation.  However this
prevents Mercury programs from dumping core.  It also prevents a parent process
from distinguishing whether a child was killed by a signal or exited with a
non-zero exit status.  This change re-throws signals that cannot be handled so
that the program can dump core and does this in a way that is compatible with
the existing signal handlers.

runtime/mercury_memory_handlers.c:
    Use MR_reset_signal() and raise() to re-throw signals that we can't handle.

runtime/mercury_signal.h:
runtime/mercury_signal.c:
    Create MR_reset_signal() that resets the handler for a signal back to the
    default (SIG_DFL).