Don't pass `--linkage static' on Linux hosts when checking for a recent mmc.
Branches: main, 10.04

Don't pass `--linkage static' on Linux hosts when checking for a recent mmc.
Some distributions no longer install static C libraries by default (e.g.
Fedora).  I also happen to be working on a Ubuntu Linux machine which was
recently upgraded to a broken version of binutils that breaks static
executables on stripping (binutils PR 10337).

The log message for the original change to force static linking says it was
needed for MSVC (for some reason).  We should consider enabling static linking
only where it is known to be required.

configure.in:
        As above.