Fix bug #153.
Branches: main, 10.04

Fix bug #153.

An empty table of contents (TOC) would be generated in the PDF version of the
user's guide.  This is caused when the user_guide.pdf make rule is executed
before the user_guide.dvi_log rule.  pdftexi executes tex only once, therefore
it is not able to generate the TOC.  Executing the .dvi_log rule before the
.pdf rule solves this problem because texinfo (used to create the .dvi) runs
tex the correct number of times.

doc/Mmakefile:
    Make %.dvi_log a dependency of %.pdf so that pdftex can use the TOC
    generated by texinfo.

doc/texinfo.tex:
    Update the version of texinfo, the existing version was from 1999.  Some
    users reported having invalid TOCs when using older versions of texinfo.