Sort out the confusion in various places between MR_bool and MR_Bool.
Branches: 10.04, main

Sort out the confusion in various places between MR_bool and MR_Bool.
MR_Bool is intended to be the C representation of the library type
bool.bool/0; MR_bool a Boolean type for use in the Mercury implementation.

In 2002, MR_Bool was deprecated, the intention apparently being to replace it
with MR_bool.  This doesn't actually work since the two are not always
interchangeable: MR_bool is a typedef for an int, while MR_Bool is word-sized -
on x86-64 for example,  the former is a 32-bits, the latter has 64-bits.  (The
mismatch can be source of confusion in foreign_procs that manipulate values of
type bool.bool/0.)

Another minor point is that if we are going to have the runtime define a
typedef for the C representation of bool.bool/0 then it should at least be
consistent with the names used for other types, e.g.

	char              -  MR_Char
	int               -  MR_Integer
	float		  -  MR_Float
	comparison_result -  MR_Comparison_Result
	bool              -  MR_Bool (rather than MR_bool)

	etc ...

This change reinstates MR_Bool (making into a typedef for MR_Word in
the process) and updates the ``C data passing conventions'' section of
the reference manual accordingly.

runtime/mercury_bootstrap.h:
	Delete the definition of MR_Bool from here.
	(This change addresses the XXX here.)

runtime/mercury_types.h:
	Put the definition of MR_Bool back here.

doc/reference_manual.texi:
	The C type corresponding to bool.bool/0 is MR_Bool not MR_bool.

library/pretty_printer.m:
	Use MR_NO rather than MR_FALSE in a spot.  (The two are identical
	but the former is preferable in the case where we are manipulating
	a value of type bool.bool/0.)

	Unrelated chage: avoid a warning concerning a pointer value being used
	as an integer.