Fix a bug in the dependant parallel conjunction transformation.
The problem occurs when a wait is pushed into a callee and that callee consumes
the future only on some of its success paths.  The caller assumes that the
callee will always wait and that it is safe to use future_get() after the call
returns.  There seems to be an invariant or assumption that the callee will
wait if it returns successfully.  To preserve this invariant the following
patch ensures that a wait is added to the end of any procedure that didn't wait
on all of its success paths.

Zoltan has suggested that it may be more ideal to put the wait on the success
paths that don't consume the variable so that in any case a variable is waited
on at most once.  This is true unless there's a non-trivial call after the
branching goal, in this case the overlap between parallel conjunctions can be
improved by waiting at the end of the procedure.

compiler/dep_par_conj.m:
    As above.