Fix a regression introduced on 2009-06-11 in the change
Branches: main, 10.04

Fix a regression introduced on 2009-06-11 in the change
"Include the type_ctor in cons_ids for user-defined types."

The symptom was an abort in the low-level C code generator when compiling
a call to an implied mode, where the implied argument is of a no-tag type.
The bug below messed up the liveness computation.

compiler/mode_util.m:
        Use `equivalent_cons_ids' to compare a cons_id-for-a-type with a
        cons_id-for-an-inst as the latter may have a dummy type_ctor field.

tests/valid/Mmakefile:
tests/valid/implied_mode_no_tag.m:
        Add test case.