Avoid a C compiler warning due to the system headers on MinGW not including a
Branches: main

Avoid a C compiler warning due to the system headers on MinGW not including a
declaration for getpagesize().  This, according to the MinGW developers, is
deliberate, despite getpagesize() being defined in libgcc on that system
it isn't supposed to be used -- this diff redirects calls to getpagesize()
on MinGW to the Win32 function GetSystemInfo().

runtime/mercury_memory.c:
	As above.

	Rename the definition of getpagesize() we use with MSVC and use
	that with MinGW as well.  Add a macro that maps calls to
	getpagesize() onto the renamed function.