Add new command line options for specifying the environment type(s) to the
Branches: main

Add new command line options for specifying the environment type(s) to the
compiler.  (By environment type, we mean the type of the shell, plus various
system utilities that may be invoked by the compiler.)

One new option, --host-env-type, specifies the environment under which the
compiler will be invoked; the other, --target-env-type, specifies the
environment under which a generated executable will be run.  (In the case,
where they are the same, a third new option, --env-type, can be used to set
both.)

On problem that this is intended to address is that when compiling Mercury
programs in the Java, Erlang or C# grades that are targeted at Windows
(e.g. to be run from the Windows command prompt), we should generate a batch
file rather than a shell script for the wrapper script.  (Moreover, in that
situation a user should still have the option of invoking mmc from the
Cygwin or MinGW shells.)

Future use of this feature will include removing some questionable assumptions
about the environment from the compiler, e.g. fixing the XXX in
options.quote_arg/1, as well improving support for cross-compilation in
general.

The environment types currently recognised are: posix, cygwin, msys and
windows.  Cygwin and MinGW/MSYS are treated separately since they are often
different enough from "posix" that this is required.  "windows"  means the
Windows command prompt, i.e. cmd.exe; although we may add another variant for
use with PowerShell in the future.

This diff adds support for the new options; once it has bootstrapped I will
post a change that fixes the problem with the Java etc grades mentioned above
and add the remainder of the documentation.

compiler/options.m:
	Add the new options: --host-env-type, --target-env-type, --env-type.

compiler/globals.m:
	Add a type describing the possible environment types.

	Add slots to the globals structure for the host and target
	environment types.  Add predicates for querying the
	environment types.

compiler/handle_options.m:
	Handle the new options.