Allow the dependant parallelization transformation to push waits/signals into other procedures in more cases.
A bug exists in some cases when signals and waits are pushed into code that
itself has a parallel conjunction or has already had signals and waits pushed
into it.  We previously disabled pushing waits and signals into such procedures
to avoid the bug, but this is too conservative.

We now allow this in three new cases:

    When the call is a recursive call.

    When the callee is an earlier version (WRT the specialisation transformation)
    of the caller.

    When the callee is already specialised and the
    specialisation-of-the-specialisation we want to create has already been
    created.

compiler/dep_par_conj.m:
    As above.