This change improves the worst-case performance of the compiler when asked to
Estimated hours taken: 1
Branches: main

This change improves the worst-case performance of the compiler when asked to
compile programs in debug grades. In the usual case, as measured by speedtest
when compiling for the debug grade, the result is a very small improvement
(0.3%), which is in the noise.

compiler/continuation_info.m:
	When generating the information for label layout structures,
	do not gather information about the instantiation states of variables.
	There is no point in doing so, since the debugger ignores this
	information, and in some cases (zm_enums.m from g12 in particular),
	the groundness checks we used to do are a nontrivial bottleneck,
	because (a) there are lots of insts, and (b) they can be huge.

	Replace some nondet code with more efficient det code.

compiler/llds.m:
	Change the representation of variables in label layout structures
	to ignore insts.

compiler/llds_out_instr.m:
compiler/trace_gen.m:
	Conform to the change in llds.m.

tools/speedtest:
	Generalize the options handling to allow the specification of any grade
	for the test.