Fix two more problems with tree_bitset.difference.
Estimated hours taken: 5
Branches: main

Fix two more problems with tree_bitset.difference.

The first problem, fixed by Peter Wang, was a stray "+1". This was in code
that tried to figure out the difference between two lists of nodes that could
be at different levels, and occurred when the level of the first operand
was lower than the level of the second.

The second problem occurred when the level of the first operand was greater
than the level of the second. The code tried to handle this case, but did not
do so correctly, mainly because it assumed that raising by one level is always
enough.

The underlying cause of both problems was that the predicate involved
(interiorlist_difference) tried to do too much: compute the difference AND
match the levels of the input operands. The overall fix is to get the caller
to ensure that the two operands are at the same level.

This diff was tested with
- set_of_var using test_bitset instead of tree_bitset, to test the results
  of every set_of_var operation,
- with tree_bitset's bits_per_level set to 1, to force the execution of code
  dealing with different levels of nodes much more frequently than the usual
  bits_per_level=5 (with the latter setting, only enormous sets ever have more
  than one level of interior node), and
- with integrity tests on tree_bitset.m enabled.

library/tree_bitset.m:
	Make the above change.

tests/hard_coded/tree_bitset_difference.{m,exp}:
	Update the test case to test all the operand pairs listed in the Mantis
	bug report, in two forms: one that showed the bug on 32-bit systems,
	and one that showed the bug on 64-bit systems. (The word size matters
	because one leaf node contains one word's worth of bits.)

	Since we now generate much more output, checking the equivalence
	of the results generated by tree_bitset.m and by set.m by eye is no
	longer trivial. Make the code do the check itself, and print output
	only if there is a mismatch.