Support new options variables in Mercury.options files that allow flags
Branches: main, 11.07

Support new options variables in Mercury.options files that allow flags
for a specific type of C compiler to be given.  The new variables are:

   GCC_FLAGS   - for C compiler flags specific to GCC.
   CLANG_FLAGS - for C compiler flags specific to clang.
   MSVC_FLAGS  - for C compiler flags specific to Visual C.

The above are only passed to the C compiler when the C compiler type
corresponds to the variable; they are ignored if it does not.  The motivation
for this addition is that you often need to pass some additional flags to a
specific C compiler, e.g. to suppress a warning or disable an optimization, but
using only mmc --make and options files there is currently no easy way to this.

The above functionality is currently only implemented for mmc --make.
(I don't have any plans to implement it for mmake.)

compiler/options_file.m:
	Support the three new variables.

compiler/compile_target_code.m:
	When collecting the flags to pass to the C compiler, also gather
	up any flags specifically for a C compiler whose type matches the
	current C compiler type.

compiler/options.m:
	Add new internal options required by the above.

doc/user_guide.texi:
	Document the above.